Lesson 2

+++ 1 lesson/triangle_area.rb
@@ -1,7 +1,7 @@
 print "Enter triangle base (a): "
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 print "Enter triangle height (h): "
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 area = 0.5 * base * height
